% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static aggregate risk development.R
\name{static_aggregate_risk_development}
\alias{static_aggregate_risk_development}
\title{Aggregate Upside and Downside Risk Over Time}
\usage{
static_aggregate_risk_development(
  file_paths_named_list,
  infl_col = c(16),
  upside_col = c(17:22),
  downside_col = c(24:29),
  xlab = "Survey Date",
  ylab = "Share of Total Risk",
  title = "Share of Aggregated Upside vs Downside Risk Over Time"
)
}
\arguments{
\item{file_paths_named_list}{A named list of file paths to Excel files. The names are used as survey labels.}

\item{infl_col}{Index of the inflation column (default is 16).}

\item{upside_col}{Vector of column indices representing upside risks (default is 17:22).}

\item{downside_col}{Vector of column indices representing downside risks (default is 24:29).}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{title}{Title of the plot.}
}
\value{
A ggplot object showing the share of aggregated upside and downside risk over time.
}
\description{
Aggregates risk perceptions from multiple survey waves and returns a stacked bar plot
of the relative share of upside vs. downside risks over time.
}
\examples{
\dontrun{
file_list <- nbssma::prepare_file_list(c=("May 25", "Jun 25", "Jul 25"))
aggregate_risk_development(file_list)
}

}
\author{
Ole Paech
}
