% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static bar line.R
\name{static_bar_line}
\alias{static_bar_line}
\title{Create a bar and line plot showing median and IQR over time (staitc version)}
\usage{
static_bar_line(
  data,
  rel_cols = c(10, 12, 14),
  title1 = "Median in \% (left)",
  title2 = "IQR (right)"
)
}
\arguments{
\item{data}{A data.frame with the survey data.}

\item{rel_cols}{A vector stating in which columns of the file the data to visualize are.}

\item{title1}{A character string specifying the title of the left y-axis (optional).}

\item{title2}{A character string specifying the title of the right y-axis (optional).}
}
\value{
A ggplot2 object with dual-axis style median (bar) and IQR (line) chart.
}
\description{
Create a bar and line plot showing median and IQR over time (staitc version)
}
\examples{
path <- nbssma::load_participant_files()
data <- readxl::read_excel(path)
static_bar_line(data, rel_cols = c(10,12,14), title1 = "Median (left)", title2 = "IQR (right)")

}
\author{
Ole Paech
}
