% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_boxplot.R
\name{static_boxplot}
\alias{static_boxplot}
\title{Static Boxplot for Median Expected DFR Rates}
\usage{
static_boxplot(
  data,
  rel_cols = c(10, 12, 14),
  xlab = "",
  ylab = "Median Rate (in \%)",
  title = ""
)
}
\arguments{
\item{data}{A data frame containing percentage values as character strings.}

\item{rel_cols}{A vector of column indices to be plotted (default: c(10, 12, 14)).}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis (default: "Median Rate (in \%)").}

\item{title}{Plot title.}
}
\value{
A \code{ggplot2} object showing boxplots for the dfr expectation.
}
\description{
This function creates a static boxplot visualization for selected inflation-related
survey columns. Percent values are cleaned, converted to numeric, and plotted by month.
}
\examples{
\dontrun{
path <- nbssma::load_participant_files()
df <- readxl::read_excel(path)
static_boxplot(data, rel_cols = c(10, 12, 14))
}

}
\author{
Ole Paech
}
