% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static boxplot categories.R
\name{static_boxplot_categories}
\alias{static_boxplot_categories}
\title{Create facetted boxplots grouped by respondent category (ggplot2 version)}
\usage{
static_boxplot_categories(
  data,
  category,
  rel_cols = c(10, 12, 14),
  xlab = "",
  ylab = "Rate (in \%)",
  title = ""
)
}
\arguments{
\item{data}{A data frame containing survey results.}

\item{category}{A string indicating the grouping category. Must be one of
\code{"Profession"}, \code{"Experience"}, or \code{"Nationality"}.}

\item{rel_cols}{A vector stating in which columns of the file the data to visualize are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}
}
\value{
A static ggplot2 object containing facetted boxplots by month and category.
}
\description{
This function takes NBS survey data and generates facetted boxplots of expectations
grouped by a respondent category (profession, experience, or nationality)
across different time periods.
}
\examples{
\dontrun{
path <- nbssma::load_participant_files()
data <- readxl::read_excel("path")
static_boxplot_categories(data, category = "Experience")
}

}
\author{
Ole Paech
}
