% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static dotplot category.R
\name{static_dotplot}
\alias{static_dotplot}
\title{Create a dotplot of rates by month and category}
\usage{
static_dotplot(
  data,
  category,
  rel_cols = c(10, 12, 14),
  xlab = "",
  ylab = "Rate (in \%)",
  title = ""
)
}
\arguments{
\item{data}{A data frame containing survey data.}

\item{category}{A string indicating the grouping category. Must be one of
\code{"Profession"}, \code{"Experience"}, or \code{"Nationality"}.}

\item{rel_cols}{A vector stating in which columns of the file the data to visualize are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}
}
\value{
A static ggplot2 plot object.
}
\description{
This function takes NBS survey data and creates a dotplot
showing individual rate observations by month, colored by a
respondent category (profession, experience, or nationality).
}
\examples{
\dontrun{
path <- nbssma::load_participant_files()
data <- readxl::read_excel("path")
static_dotplot(data, category = "Experience")
}

}
\author{
Ole Paech
}
