% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static bars category.R
\name{static_group_bar_category}
\alias{static_group_bar_category}
\title{Creates a grouped bar chart showing Median, Mean, and Mode of selected columns
grouped by a specified category (e.g., Profession, Experience, or Nationality).}
\usage{
static_group_bar_category(
  data,
  category,
  rel_cols = c(10, 12, 14),
  ylab = "Median Rate (in \%)",
  title = ""
)
}
\arguments{
\item{data}{A data frame containing categorical and numeric columns.}

\item{category}{A string: "Profession", "Experience", or "Nationality".}

\item{rel_cols}{Integer indices of the relevant numeric columns (default: c(10, 12, 14)).}

\item{ylab}{Label for the y-axis.}

\item{title}{Title for the plot.}
}
\value{
A ggplot object showing the grouped bar chart by category and month.
}
\description{
Creates a grouped bar chart showing Median, Mean, and Mode of selected columns
grouped by a specified category (e.g., Profession, Experience, or Nationality).
}
\examples{
\dontrun{
path <- nbssma::load_participant_files()
data <- readxl::read_excel(path)
static_group_bar_category(data, category = "Profession")
}
}
\author{
Ole Paech
}
