% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static grouped Historgram.R
\name{static_grouped_histogram}
\alias{static_grouped_histogram}
\title{Grouped Histogram Plot of Expected Rates by Month (Static Version)}
\usage{
static_grouped_histogram(
  data,
  rel_cols = c(10, 12, 14),
  xlab = "Expected Rate",
  ylab = "Percentage (\%)",
  title = ""
)
}
\arguments{
\item{data}{Data frame containing the survey responses.}

\item{rel_cols}{A vector stating in which columns of the file the data to visualize are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}
}
\value{
A static ggplot2 histogram plot.
}
\description{
Creates grouped histograms of expected rates over months with medians highlighted.
}
\examples{
\dontrun{
path <- nbssma::load_participant_files()
data <- readxl::read_excel("path")
static_grouped_histogram(data)
}

}
\author{
Ole Paech
}
