% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static inflation history.R
\name{static_inflation_risk_history}
\alias{static_inflation_risk_history}
\title{Visualize Historical Inflation Risk Perception (Static Plot)}
\usage{
static_inflation_risk_history(
  files_with_labels,
  infl_col = c(16),
  upside_col = c(17:22),
  downside_col = c(24:29),
  xlab = "Survey Date",
  ylab = "Average Inflation Expectations (in \%)",
  title = "Development of Inflation Projections and Perceived Risks"
)
}
\arguments{
\item{files_with_labels}{A chronological named list, where each element is the path
to an Excel file and the name is the corresponding survey date label.}

\item{infl_col}{A vector stating in which column of the file the inflation data are.}

\item{upside_col}{A vector stating in which columns of the file the upside risk data are.}

\item{downside_col}{A vector stating in which columns of the file the downside risk data are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}
}
\value{
A static ggplot2 object showing inflation expectations and risk decomposition.
}
\description{
This function processes labeled Excel survey files and visualizes
the historical development of upside and downside inflation risk
in relation to inflation expectations using a stacked bar plot.
}
\examples{
\dontrun{
file_list <- nbssma::prepare_file_list(c=("May 25", "Jun 25", "Jul 25"))
static_inflation_risk_history(file_list)
}

}
\author{
Ole Paech
}
