% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static risk factors development.R
\name{static_risk_share_bars}
\alias{static_risk_share_bars}
\title{Create a Risk Share Bar-plot (Upside or Downside)}
\usage{
static_risk_share_bars(
  file_paths_named_list,
  type = "Upside",
  infl_col = c(16),
  upside_col = c(17:22),
  downside_col = c(24:29),
  xlab = "Survey Date"
)
}
\arguments{
\item{file_paths_named_list}{Named list of Excel file paths.}

\item{type}{Character string specifying risk type: either "Upside" (default) or "Downside".}

\item{infl_col}{A vector stating in which column of the file the inflation data are.}

\item{upside_col}{A vector stating in which columns of the file the upside risk data are.}

\item{downside_col}{A vector stating in which columns of the file the downside risk data are.}

\item{xlab}{A character string specifying the x-axis label (optional).}
}
\value{
A ggplot2 bar plot showing risk share distribution over time.
}
\description{
This function reads SMA survey data from multiple files and visualizes the composition of
upside or downside risk over time as a stacked bar plot.
}
\author{
Ole Paech
}
